# ENH: Move `loadtxt` to C for much better speed