--- a/numpy/core/src/multiarray/textreading/growth.h
+++ b/numpy/core/src/multiarray/textreading/growth.h
@@ -1,0 +1,7 @@
+#ifndef _NPY_GROWTH_H
+#define _NPY_GROWTH_H
+
+npy_intp
+grow_size_and_multiply(size_t *size, size_t min_grow, npy_intp itemsize);
+
+#endif  /*_NPY_GROWTH_H */