--- a/numpy/core/src/multiarray/textreading/readtext.h
+++ b/numpy/core/src/multiarray/textreading/readtext.h
@@ -1,0 +1,7 @@
+#ifndef READTEXT_H_
+#define READTEXT_H_
+
+NPY_NO_EXPORT PyObject *
+_load_from_filelike(PyObject *self, PyObject *args, PyObject *kwargs);
+
+#endif  /* READTEXT_H_ */