… doc/release/upcoming_changes/20580.compatibility.rst
… doc/release/upcoming_changes/20580.new_feature.rst
… doc/release/upcoming_changes/20580.performance.rst
… numpy/core/src/multiarray/textreading/growth.c
… numpy/core/src/multiarray/textreading/parser_config.h
… numpy/core/src/multiarray/textreading/readtext.c
… numpy/core/src/multiarray/textreading/rows.c
… numpy/core/src/multiarray/textreading/rows.h
… numpy/core/src/multiarray/textreading/str_to_int.c
… numpy/core/src/multiarray/textreading/str_to_int.h
… numpy/core/src/multiarray/textreading/stream.h
… numpy/core/src/multiarray/textreading/stream_pyobject.c
… numpy/core/src/multiarray/textreading/stream_pyobject.h
… numpy/core/src/multiarray/textreading/tokenize.c.src
… numpy/core/src/multiarray/textreading/tokenize.h
… numpy/lib/npyio.py
… numpy/lib/tests/test_io.py
… numpy/lib/tests/test_loadtxt.py