This PR is based on @WarrenWeckesser `npreadtext` which was subsequently heavily modified [here](https://github.com/BIDS-numpy/npreadtext)  (and somewhat reduced in scope).

While there is more to it, the main point is to improve the speed for parsing CSV (like) files, the included io benchmark change the following way:

<details> <summary><b>Benchmarking results</b>  </summary>

(sorted so that the `ratio` is a factor, 17 means 17 times faster)
EDIT: Last table update Jan 14 (no meaningful changes expected anymore) 
```
       before           after         ratio
     [1e6b72b4]       [e2d35064]
     <add-npreadtext>       <main>    
+     9.58±0.07ms          178±2ms    18.64  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100000)
+         949±5μs       17.5±0.1ms    18.42  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 10000)
+      8.30±0.1ms        142±0.1ms    17.16  bench_io.LoadtxtCSVStructured.time_loadtxt_csv_struct_dtype
+      9.11±0.1ms        131±0.2ms    14.40  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100000)
+        918±10μs       13.1±0.1ms    14.24  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10000)
+     13.4±0.08μs          188±1μs    14.05  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100)
+     13.1±0.05μs          143±1μs    10.91  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100)
+      15.6±0.4ms          135±1ms     8.67  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100000)
+      14.6±0.1ms          127±1ms     8.67  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100000)
+      14.5±0.1ms          125±1ms     8.58  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100000)
+     1.45±0.01ms      12.3±0.09ms     8.49  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10000)
+     15.1±0.07ms        126±0.7ms     8.40  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100000)
+     1.48±0.02ms      12.4±0.03ms     8.39  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10000)
+     1.60±0.01ms       13.4±0.1ms     8.38  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10000)
+        1.52±0ms      12.5±0.05ms     8.23  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10000)
+     19.1±0.03μs        146±0.2μs     7.63  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100)
+      18.4±0.2μs        135±0.9μs     7.33  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100)
+     18.7±0.08μs        137±0.5μs     7.30  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100)
+     18.9±0.06μs          137±2μs     7.21  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100)
+     1.55±0.01ms      10.3±0.05ms     6.66  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10000)
+     16.0±0.08ms          105±1ms     6.60  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100000)
+        916±10μs      5.98±0.02ms     6.52  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3])
+     1.15±0.02ms      7.36±0.01ms     6.40  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3, 5, 7])
+      18.2±0.3μs          110±1μs     6.03  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100)
+        798±10μs      4.80±0.01ms     6.02  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv(2)
+     5.07±0.01μs       29.7±0.7μs     5.86  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 10)
+     7.36±0.08ms       39.7±0.2ms     5.39  bench_io.LoadtxtCSVDateTime.time_loadtxt_csv_datetime(20000)
+      77.1±0.5μs          413±3μs     5.35  bench_io.LoadtxtCSVDateTime.time_loadtxt_csv_datetime(200)
+         737±7μs      3.92±0.01ms     5.32  bench_io.LoadtxtCSVDateTime.time_loadtxt_csv_datetime(2000)
+     5.06±0.02μs       25.1±0.5μs     4.95  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10)
+     12.3±0.02μs       58.3±0.5μs     4.74  bench_io.LoadtxtCSVDateTime.time_loadtxt_csv_datetime(20)
+     5.67±0.03μs       26.3±0.4μs     4.64  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10)
+         220±4μs          989±4μs     4.49  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(1000)
+         221±5μs          987±8μs     4.46  bench_io.LoadtxtReadUint64Integers.time_read_uint64(1000)
+         123±3μs          549±2μs     4.46  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(550)
+         124±1μs          548±2μs     4.41  bench_io.LoadtxtReadUint64Integers.time_read_uint64(550)
+     2.22±0.03ms      9.76±0.04ms     4.41  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(10000)
+     2.22±0.03ms      9.69±0.04ms     4.37  bench_io.LoadtxtReadUint64Integers.time_read_uint64(10000)
+     5.62±0.01μs       24.2±0.8μs     4.31  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10)
+     5.64±0.01μs       24.0±0.1μs     4.26  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10)
+     5.44±0.03μs       22.8±0.3μs     4.20  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10)
+     5.90±0.02μs       21.8±0.6μs     3.69  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10)
+     4.22±0.01ms      12.7±0.08ms     3.02  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
+      46.6±0.3μs        137±0.7μs     2.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100)
+      42.6±0.1ms        122±0.7ms     2.87  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100000)
+      73.1±0.7ms        202±0.4ms     2.76  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(0)
+      72.6±0.7ms        200±0.7ms     2.76  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(500)
+      66.1±0.9ms        182±0.6ms     2.76  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(10000)
+      10.2±0.2μs       24.3±0.1μs     2.39  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10)
```

</details>

**Marking as draft, but I do not expect major changes, so this can get a first round of review, I think.**

There is a small list of TODOs:
* [x] The current additional tests from npreadtext are missing here (they should be rewritten in terms of loadtxt probably).
* [x] Additional tests should be added (see also npreadtext issue tracker;  one reason for the PR, it makes checking code-coverage more convenient.)
* [x] Smaller changes ~(e.g. I think I want to remove passing `usecols` as a numpy array)~ Should be mostly done. ~Moving usecols could be done (and would be nice), but we don't have quite the right stuff available so added the correct sanity checks instead.  I agree this is a wart (a private one), so if anyone insists I can look at it.~ Moved usecols as well, nothing is quite awesome, but I think it is _better_.

The [issues from `nreadtext`](https://github.com/BIDS-numpy/npreadtext/issues) still apply, and we should review some of these or make a call on them.  The issue tracker there may also be a good way to open other smaller things to fix-up (to reduce noise here).

---

As a general overview, the parser has a tokenizer, that works with unicode, but supports the Python modes UCS4, UCS2, and UCS1.  It parses everything into smaller UCS4 (unicode) strings, this copies the data – in theory not always necessary, but convenient – although future optimization could be done.
After this step, we have dtype (especially for integers) specific converter functions which directly convert the UCS4 string into the final result.  If necessary, we fall back to `PyArray_Pack`.
(Since we do not discover the result `dtype`, strings have to offload some work to Python currently.)

In general, we read either line-by-line from Python iterables, or in chunks if reading from files.

The new features are currently fairly limited (e.g. I removed the custom float parser to support using `d` instead of `e`), and is not fully CVS parser featured (but this can easily be achieved now!).
One particular important changes is the current decision to remove hex float parsing and limit integer parsing to integers (we used to fall back to floats).  The custom parsers also mean that we do not support numbers with an underscore `100_000` right now.
Our current idea was that we should improve the error message a bit more, but generally advise users to use `converters=int` in this case.

Please check the `npreadtext` issue tracker for issues/changes: https://github.com/BIDS-numpy/npreadtext/issues

EDIT: One big additional feature ~(which I have not made public here, but is as easy as adding the kwarg)~, is that the parser does support quoting (with excel double-quote escaping).