@seberg 2021-12-13
(The PyPy test failure should be the same as in the `empty`/intp sequence parsing PR: PyPy had a bug with formatting `%(100)R`, it is fixed, and for now we should just ignore it.)