@bashtage 2021-12-14
Looking at [this post](https://juliacomputing.com/blog/2020/06/fast-csv/) it seems to me that unless one wants to do a multithreaded reader then the best case gains are pretty limited (if one wants to compare to Julia as a reference point).