@seberg 2021-12-14
> 3k lines for a 50% improvement over pandas when using maximum float precision seems like a less-than-ideal ratio (and slower if one is happy with the default).

Yes, IIRC, compared to pandas this has basically parity or up to 50% faster.  But, I think you will find the C code to be much more maintainable and much easier to extend.  Some notes:
* To get feature-parity with loadtxt, the pandas version must use its Python backend which is _also_ extremely complex (and very slow of course).  This is for things like:
  * Reading any file that isn't a raw utf-8 file.
  * Use any control character that is not ASCII.  E.g. splitting by whitespace `delimiter=None` doesn't work, pandas only supports ascii whitespace with any decent speed (even with `delimiter=','` non-ascii whitespace will not get stripped when other whitespace does get stripped).
  On the contrary, here the reading itself is still largely using Python API to keep it simple and flexible (with a line-by-line vs. chunked mode adding a bit of complexity for ~30% speedup, I admit).  Meaning that all of those cases are covered by the same code.
* I am very sure that if you will have much more problems extending the pandas code.  I wrote this in a way that it should be straight forward to e.g. allow users to supply parsers written in C directly, or allow swapping out our parsers (if you want to have that fast float parser). I like the idea to make it so that users can supply `converters=faster_float_parser_written_in_C`.  (yes, we would have to nail down public API for that, but the general design is there.)
* The pandas code is much less maintainable also IMO.  It has as many lines of code, but is not nearly as well structured.

Is Cython better?  I guess opinions will differ on that.  It is probably more accessible to contributions, but honestly, I am not sure I like Cython for code that is "heavy C".  It is great for the interface code, but the interface code isn't the big chunk here.