@mattip 2021-12-14
What are the benchmarks for the code? Is there something that exercises the non-float dtypes?