@seberg 2021-12-24
> The test suite changes look fairly modest, and include some improved error messages it seems.

Yeah, the main changes I am aware of are:
* more restrictive parsing: no floats when storing integers, no underscore `100_000` support (can be gotten back using `converters=float` although a bit annoying if you try to use both versions, since the old one requires listing all columns.
* There is some strange stuff about counting lines vs. counting rows: gh-19785 (this can be fixed in any way pretty much, but needs a decision).  Also if you pass in rows that contain newlines, things get a bit weird.  I did not care much, because it seems super strange, but it is a change

The other things should really be improvements, like the better errors (although sometimes changed) – could still add error chaining there, especially for user provided converters.

In any case, thanks for taking a look!  I will be off mostly for the next 1.5 weeks, so probably no updates :).